requirements.txt
python
django
numpy
tensorflow
cv2
torch
sklearn
pillow
